refactor(api): move github to modules/ and split its webhook from its settings
