+++ apps/api/src/app.ts
@@ -15,7 +15,7 @@ import { setMcpApp } from './mcp/app-ref';
 import { internalAgentRunRoutes } from './modules/agents/core/internal-routes';
 import { internalNotificationRoutes } from './notifications/internal-routes';
 import { internalTelegramRoutes } from './telegram/internal-routes';
-import { githubWebhookRoutes } from './github/routes';
+import { githubWebhookRoutes } from './modules/github/webhook';
 
 // The assembled Elysia app, without `.listen()`. `index.ts` imports this and
 // binds the port; tests import it and pass it to Eden Treaty to drive routes in
