+++ apps/api/src/modules/github/magic-words.ts
